python_api/requirements.txt
flask
flask-cors
pillow
pytesseract
opencv-python-headless
tensorflow
numpy